tasks/generative-ai/text-to-text/fine-tuning/instruction-tuning/RWKV/scripts/code/requirements.txt
transformers
pytorch-lightning==1.9
deepspeed
wandb
ninja
pynvml==11.4.1
rwkv
